plotly_basic/requirements.txt
plotly
cufflinks
numpy
geopandas
pyshp
shapely
dash==0.21.0  # The core dash backend
dash-renderer==0.12.1  # The dash front-end
dash-html-components==0.10.0  # HTML components
dash-core-components==0.22.1  # Supercharged components
